Add files via upload
I added the newly made Money Puck recommender engine and the functions to call when using it. In the upload it includes the MoneyPuck processing pipeline which uses the standard scaler and PCA. Before doing that I needed to fix the inf and NaN values that occurring in my avg_ice_time/shift column.